agriculture-rag-react/backend/requirements.txt
flask
flask-cors
werkzeug
langchain
langchain-community
langchain-openai
langchain-core
pypdf
pdf2image
pillow
python-dotenv
openai
sentence-transformers
faiss-cpu
transformers
torch
tiktoken
pymupdf
numpy
tqdm